Fix session problem in Cart

+++ backend/cart/views.py
@@ -6,16 +6,11 @@ from django.utils import translation
 from product.models import Product
 from rest_framework import status
 from django.db.models import Sum
+from django.conf import settings
 
 
 def get_main_image_url(request, product):
-    """Метод для отримання абсолютного URL головного зображення товару.
-
-    Args:
-        request (HttpRequest): Поточний запит.
-        product (Product): Об'єкт товару.
-
-    """
+    """Повертає абсолютний URL головного зображення товару."""
     main_image = product.productimage_set.filter(is_main=True).first()
     if main_image and main_image.image:
         return request.build_absolute_uri(main_image.image.url)
@@ -42,7 +37,11 @@ class CartViewSet(ViewSet):
 
     permission_classes = [AllowAny]
 
-    def get_cart(self, request) -> Cart:
+    def activate_translation(self, request) -> None:
+        language = request.headers.get("Accept-Language", "en")
+        translation.activate(language)
+
+    def get_cart(self, request):
         """
         Повертає кошик користувача або створює його, якщо відсутній.
 
@@ -54,20 +53,11 @@ class CartViewSet(ViewSet):
         """
         if request.user.is_authenticated:
             cart, created = Cart.objects.get_or_create(user=request.user)
+            return cart
         else:
-            session_key = request.session.session_key
-            if not session_key:
-                request.session.create()
-                session_key = request.session.session_key
-
-            cart, created = Cart.objects.get_or_create(session_id=session_key)
-
-        return cart
-
-    def activate_translation(self, request) -> None:
-        """Активує переклад на основі заголовку 'Accept-Language'."""
-        language = request.headers.get("Accept-Language", "en")
-        translation.activate(language)
+            if settings.CART_SESSION_ID not in request.session:
+                request.session[settings.CART_SESSION_ID] = {}
+            return request.session[settings.CART_SESSION_ID]
 
     def list(self, request) -> Response:
         """
@@ -79,15 +69,12 @@ class CartViewSet(ViewSet):
 
         """
         self.activate_translation(request)
-        cart = self.get_cart(request)
-        cart_items = CartItem.objects.filter(cart=cart)
-        total_price = cart_items.aggregate(total=Sum("price"))["total"] or 0
+        if request.user.is_authenticated:
+            cart = self.get_cart(request)
+            cart_items = CartItem.objects.filter(cart=cart)
+            total_price = cart_items.aggregate(total=Sum("price"))["total"] or 0
 
-        response_data = {
-            "session": request.session.session_key,
-            "cart_id": cart.id,
-            "total_price": total_price,
-            "items": [
+            items = [
                 {
                     "product_id": item.product.id,
                     "product_name": item.product.name,
@@ -96,8 +83,36 @@ class CartViewSet(ViewSet):
                     "main_image_url": get_main_image_url(request, item.product),
                 }
                 for item in cart_items
-            ],
-        }
+            ]
+            response_data = {
+                "cart_id": cart.id,
+                "total_price": total_price,
+                "items": items,
+            }
+        else:
+            session_cart = self.get_cart(request)
+            items = []
+            total_price = 0
+            for product_id_str, data in session_cart.items():
+                try:
+                    product = Product.objects.get(pk=int(product_id_str))
+                    items.append(
+                        {
+                            "product_id": product.id,
+                            "product_name": product.name,
+                            "quantity": data["quantity"],
+                            "price": data["price"],
+                            "main_image_url": get_main_image_url(request, product),
+                        }
+                    )
+                    total_price += data["price"] * data["quantity"]
+                except Product.DoesNotExist:
+                    continue
+            response_data = {
+                "cart_id": None,
+                "total_price": total_price,
+                "items": items,
+            }
 
         return Response(response_data, status=status.HTTP_200_OK)
 
@@ -115,16 +130,13 @@ class CartViewSet(ViewSet):
                 - 400 BAD REQUEST: Якщо передано некоректну кількість товару.
                 - 404 NOT FOUND: Якщо товар не знайдено.
         """
-        cart = self.get_cart(request)
         product_id = request.data.get("product_id")
         quantity = request.data.get("quantity", 1)
-
         if not isinstance(quantity, int) or quantity <= 0:
             return Response(
                 {"error": "Quantity must be a positive integer."},
                 status=status.HTTP_400_BAD_REQUEST,
             )
-
         try:
             product = Product.objects.get(pk=product_id)
         except Product.DoesNotExist:
@@ -132,22 +144,39 @@ class CartViewSet(ViewSet):
                 {"error": "Product not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
-        cart_item, created = CartItem.objects.get_or_create(
-            cart=cart,
-            product=product,
-            defaults={"quantity": quantity, "price": product.price},
-        )
-
-        if not created:
-            cart_item.quantity += quantity
-            cart_item.save()
+        if request.user.is_authenticated:
+            cart = self.get_cart(request)
+            cart_item, created = CartItem.objects.get_or_create(
+                cart=cart,
+                product=product,
+                defaults={"quantity": quantity, "price": product.price},
+            )
+            if not created:
+                cart_item.quantity += quantity
+                cart_item.save()
+                message = "Product quantity updated"
+            else:
+                message = "Product added to cart"
             return Response(
-                {"message": "Product quantity updated", "quantity": cart_item.quantity},
-                status=status.HTTP_200_OK,
+                {"message": message, "quantity": cart_item.quantity},
+                status=status.HTTP_201_CREATED,
             )
         else:
+            session_cart = self.get_cart(request)
+            product_key = str(product_id)
+            if product_key in session_cart:
+                session_cart[product_key]["quantity"] += quantity
+            else:
+                session_cart[product_key] = {
+                    "quantity": quantity,
+                    "price": product.price,
+                }
+            request.session[settings.CART_SESSION_ID] = session_cart
             return Response(
-                {"message": "Product added to cart", "quantity": cart_item.quantity},
+                {
+                    "message": "Product added to cart",
+                    "quantity": session_cart[product_key]["quantity"],
+                },
                 status=status.HTTP_201_CREATED,
             )
 
@@ -166,36 +195,53 @@ class CartViewSet(ViewSet):
                 - 400 BAD REQUEST: Якщо кількість передана некоректно.
                 - 404 NOT FOUND: Якщо товар не знайдено у кошику.
         """
-        cart = self.get_cart(request)
-
-        try:
-            cart_item = CartItem.objects.get(cart=cart, product_id=pk)
-        except CartItem.DoesNotExist:
-            return Response(
-                {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
-            )
-
         change = request.data.get("quantity", 0)
-
         if not isinstance(change, int):
             return Response(
                 {"error": "Quantity must be an integer."},
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
-        cart_item.quantity += change
+        if request.user.is_authenticated:
+            cart = self.get_cart(request)
+            try:
+                cart_item = CartItem.objects.get(cart=cart, product_id=pk)
+            except CartItem.DoesNotExist:
+                return Response(
+                    {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
+                )
 
-        if cart_item.quantity <= 0:
-            cart_item.delete()
-            return Response(
-                {"message": "Product removed from cart"}, status=status.HTTP_200_OK
-            )
+            cart_item.quantity += change
+            if cart_item.quantity <= 0:
+                cart_item.delete()
+                return Response(
+                    {"message": "Product removed from cart"}, status=status.HTTP_200_OK
+                )
+            else:
+                cart_item.save()
+                return Response(
+                    {
+                        "message": "Product quantity updated",
+                        "quantity": cart_item.quantity,
+                    },
+                    status=status.HTTP_200_OK,
+                )
         else:
-            cart_item.save()
-            return Response(
-                {"message": "Product quantity updated", "quantity": cart_item.quantity},
-                status=status.HTTP_200_OK,
-            )
+            session_cart = self.get_cart(request)
+            product_key = str(pk)
+            if product_key not in session_cart:
+                return Response(
+                    {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
+                )
+
+            session_cart[product_key]["quantity"] += change
+            if session_cart[product_key]["quantity"] <= 0:
+                del session_cart[product_key]
+                message = "Product removed from cart"
+            else:
+                message = "Product quantity updated"
+            request.session[settings.CART_SESSION_ID] = session_cart
+            return Response({"message": message}, status=status.HTTP_200_OK)
 
     def destroy(self, request, pk=None) -> Response:
         """
@@ -210,27 +256,28 @@ class CartViewSet(ViewSet):
                 - 404 NOT FOUND: Якщо товар не існує або його немає у кошику.
                 - 500 INTERNAL SERVER ERROR: Якщо сталася непередбачена помилка.
         """
-        cart = self.get_cart(request)
-
-        try:
-            product = Product.objects.get(pk=pk)
-        except Product.DoesNotExist:
-            return Response(
-                {"error": "Product not found"}, status=status.HTTP_404_NOT_FOUND
-            )
-
-        try:
-            cart_item = CartItem.objects.get(cart=cart, product=product)
-            cart_item.delete()
-            return Response(
-                {"message": "Product removed from cart"}, status=status.HTTP_200_OK
-            )
-        except CartItem.DoesNotExist:
-            return Response(
-                {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
-            )
-        except Exception as e:
-            return Response(
-                {"error": f"An error occurred: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
+        if request.user.is_authenticated:
+            cart = self.get_cart(request)
+            try:
+                cart_item = CartItem.objects.get(cart=cart, product_id=pk)
+                cart_item.delete()
+                return Response(
+                    {"message": "Product removed from cart"}, status=status.HTTP_200_OK
+                )
+            except CartItem.DoesNotExist:
+                return Response(
+                    {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
+                )
+        else:
+            session_cart = self.get_cart(request)
+            product_key = str(pk)
+            if product_key in session_cart:
+                del session_cart[product_key]
+                request.session[settings.CART_SESSION_ID] = session_cart
+                return Response(
+                    {"message": "Product removed from cart"}, status=status.HTTP_200_OK
+                )
+            else:
+                return Response(
+                    {"error": "Product not in cart"}, status=status.HTTP_404_NOT_FOUND
+                )
